# jekyll: fix sitemap lastmod