--- a/.dockerignore
+++ b/.dockerignore
@@ -1,5 +1,4 @@
 .DS_Store
-.git
 .github
 .gitignore
 .idea