--- a/Gemfile
+++ b/Gemfile
@@ -16,8 +16,8 @@ end
 #  more info: https://github.com/docker/docs/issues/14788
 gem 'rouge', '3.27.0'
 
-gem 'archive-zip', '0.12.0'
 gem 'front_matter_parser', '1.0.1'
+gem 'git', '1.12.0'
 gem 'html-proofer', '3.19.4'
 gem 'mdl', '0.11.0'
 gem 'octopress-hooks', '2.6.2'