follow-up https://github.com/docker/docker.github.io/pull/15250#issuecomment-1197924756

fix sitemap last modification date not being taken into account:

![image](https://user-images.githubusercontent.com/1951866/181793200-5346ff2e-1c10-471b-a2ae-c8a34c306d14.png)

it will use git and fall back to file's `mtime` if necessary.

also adds following entries to sitemap exclusion list that should not be there: https://docs-stage.docker.com/sitemap.xml

![image](https://user-images.githubusercontent.com/1951866/181793551-f34659b2-471f-40d9-adfb-d371481f162b.png)

Signed-off-by: CrazyMax <crazy-max@users.noreply.github.com>