@crazy-max 2022-08-31
@thaJeztah Except your concerns with `.git` folder being added, it LGTY?

I made some comparison in our CI to see if there is a huge perf regression:

Last build on master branch: https://github.com/docker/docker.github.io/runs/8111429336?check_suite_focus=true#step:4:571

```
#15 45.91                     done in 45.237 seconds.
#15 45.91  Auto-regeneration: disabled. Use --watch to enable.
#15 DONE 46.3s
```

This PR: https://github.com/docker/docker.github.io/runs/8112641263?check_suite_focus=true#step:4:601

```
#15 65.50                     done in 64.467 seconds.
#15 65.50  Auto-regeneration: disabled. Use --watch to enable.
#15 DONE 66.0s
```

It seems acceptable to me and worth the change to fix our sitemap so search engines can consume it and make relevant decision on results.