@crazy-max 2022-10-26
Made some changes related to https://github.com/docker/docs/pull/15239#issuecomment-1200164942 so now remote resources are cached and fetched using Git so incremental builds are improved. This will speedup local builds by almost 55% while still generating the sitemap with last modification date:

Before:

```
#15 57.14                     done in 56.239 seconds.
#15 57.14  Auto-regeneration: disabled. Use --watch to enable.
#15 DONE 57.7s
```

Now:

```
#15 26.39                     done in 25.704 seconds.
#15 26.39  Auto-regeneration: disabled. Use --watch to enable.
#15 DONE 26.5s
```

With this new way to fetch remote resources we are now able to set the right last modification date for remote pages. cc @usha-mandya @dvdksn