@dvdksn 2022-10-26
The incremental builds are really fast with this 🤩 

But I noticed that the first build can take a really long time. Build time on `main` for a clean build is ~180 seconds for me, whereas in this branch it finished after ~240 seconds. I wonder if it's because we're cloning history? I tried setting `git.clone(depth: 1)` and it seems to fix the speed issue; I'm now back at ~180 seconds for a clean build. WDYT?

```diff
diff --git a/_plugins/fetch_remote.rb b/_plugins/fetch_remote.rb
index 4370baaccb..158fb3919a 100644
--- a/_plugins/fetch_remote.rb
+++ b/_plugins/fetch_remote.rb
@@ -55,11 +55,11 @@ module Jekyll
           rescue => e
             FileUtils.rm_rf(clonedir)
             puts "    Cloning repository into #{clonedir}"
-            git = Git.clone("#{entry['repo']}.git", Pathname.new(clonedir), branch: entry['ref'])
+            git = Git.clone("#{entry['repo']}.git", Pathname.new(clonedir), branch: entry['ref'], depth: 1)
           end
         else
           puts "    Cloning repository into #{clonedir}"
-          git = Git.clone("#{entry['repo']}.git", Pathname.new(clonedir), branch: entry['ref'])
+          git = Git.clone("#{entry['repo']}.git", Pathname.new(clonedir), branch: entry['ref'], depth: 1)
         end
 
         entry['paths'].each do |path|
```