@crazy-max 2022-10-26
> I tried setting `git.clone(depth: 1)` and it seems to fix the speed issue; I'm now back at ~180 seconds for a clean build. WDYT?

Ah good point yes we don't need the history anyway, will make this change :+1: